Give modules in arrow/python more descriptive names
